* tp/Texinfo/XS/convert/call_html_perl_function.c (call_stage_handler), tp/Texinfo/XS/convert/convert_html.c (html_stage_handler_stage_type_names, run_stage_handlers), tp/Texinfo/XS/convert/get_html_perl_info.c (html_converter_initialize_sv), tp/Texinfo/XS/main/converter_types.h (enum html_stage_handler_stage_type, HTML_STAGE_HANDLER_INFO) (HTML_STAGE_HANDLER_INFO_LIST, CONVERTER): get stage handlers from Perl and put them in HTML_STAGE_HANDLER_INFO_LIST structure in converter.  Add run_stage_handlers and call_stage_handler to call Perl registered stage handlers references.

+++ tp/Texinfo/XS/main/get_perl_info.c
@@ -281,7 +281,7 @@ void
 add_svav_to_string_list (const SV *sv, STRING_LIST *string_list,
                          enum sv_string_type type)
 {
-  int i;
+  size_t i;
   SSize_t strings_nr;
 
   dTHX;
@@ -293,7 +293,7 @@ add_svav_to_string_list (const SV *sv, STRING_LIST *string_list,
   strings_nr = av_top_index (av) +1;
   for (i = 0; i < strings_nr; i++)
     {
-      SV** string_sv = av_fetch (av, i, 0);
+      SV **string_sv = av_fetch (av, i, 0);
       if (string_sv)
         {
           const char *string;
